requirements.txt
PyYAML==3.11
cached-property==1.2.0
docker-py==1.10.3
dockerpty==0.4.1
docopt==0.6.1
enum34==1.0.4
jsonschema==2.5.1
texttable==0.8.4
future==0.15.0
requests==2.7.0
six==1.9.0
websocket-client==0.37.0
docker-compose==1.6.0
python-dockercloud==1.0.5
gevent==1.1.1